+++ CHANGELOG.md
@@ -3,4 +3,6 @@
     Corrección:
     * El botón de eliminar SK de stripe no funcionaba
     * Cambio la clase AppSettings del core por la clase Tools::settings()
+    * Se agrega opción para poder mostrar el id de cliente de stripe como línea sin coste en la factura.
+